add logFile flag to change the default log file path
